+++ var/www/html/torrentwatch-xa/templates/favorites_info.php
@@ -5,13 +5,13 @@ if (isset($style)) {
     echo $style;
 }
 ?>>
-    <input type="hidden" name="idx" id="<?php echo 'idx_' . $key; ?>" value="<?php echo $key; ?>">
+    <input type="hidden" name="idx" id="favorite_<?php echo $key; ?>_idx" value="<?php echo $key; ?>">
     <div class="favorite_name">
         <div class="left">
             <label class="item" title="Name of the Favorite">Name:</label>
         </div>
         <div class="right">
-            <input type="text" class="text" name="name" value="<?php echo $item['Name']; ?>">
+            <input type="text" class="text" name="name" id="favorite_<?php echo $key; ?>_name" value="<?php echo $item['Name']; ?>">
         </div>
     </div>
     <div class="favorite_filter">
@@ -19,7 +19,7 @@ if (isset($style)) {
             <label class="item" title="In RegEx Matching Style, use a PCRE Unicode regex such as .* to match all">Filter:</label>
         </div>
         <div class="right">
-            <input type="text" class="text" name="filter" value="<?php echo $item['Filter']; ?>">
+            <input type="text" class="text" name="filter" id="favorite_<?php echo $key; ?>_filter" value="<?php echo $item['Filter']; ?>">
         </div>
     </div>
     <div class="favorite_not">
@@ -27,7 +27,7 @@ if (isset($style)) {
             <label class="item" title="In RegEx Matching Style, use a PCRE Unicode regex to match items you don't want">Not:</label>
         </div>
         <div class="right">
-            <input type="text" class="text" name="not" value="<?php echo $item['Not']; ?>"
+            <input type="text" class="text" name="not" id="favorite_<?php echo $key; ?>_not" value="<?php echo $item['Not']; ?>"
                    title="Don't match titles with these words. You can add more than one word, separated by spaces">
         </div>
     </div>
@@ -36,7 +36,7 @@ if (isset($style)) {
             <label class="item" title="Download Directory; overrides global Download Directory">Download Dir:</label>
         </div>
         <div class="right">
-            <input type="text" class="text" name="downloaddir" value="<?php echo $item['Download Dir']; ?>">
+            <input type="text" class="text" name="downloaddir" id="favorite_<?php echo $key; ?>_downloaddir" value="<?php echo $item['Download Dir']; ?>">
         </div>
     </div>
     <?php if ($config_values['Settings']['Client'] !== "folder" && $config_values['Settings']['Also Save Torrent Files']) { ?><div class="favorite_alsosavedir" id="favorite_alsosavedir">
@@ -44,7 +44,7 @@ if (isset($style)) {
                 <label class="item" title="Also Save Directory; overrides global default">Also Save Dir:</label>
             </div>
             <div class="right">
-                <input type="text" class="text" name="alsosavedir" value="<?php echo $item['Also Save Dir']; ?>">
+                <input type="text" class="text" name="alsosavedir" id="favorite_<?php echo $key; ?>_alsosavedir" value="<?php echo $item['Also Save Dir']; ?>">
             </div>
         </div><?php } ?>
     <div class="favorite_episodes">
@@ -52,7 +52,7 @@ if (isset($style)) {
             <label class="item" title="Episode filter. ex.: 1x1-3x24 for Season 1 Episode 1 to Season 3 Episode 24. To just set a starting point use: 2x10. You may use s01e12 instead of 1x12." >Episodes:</label>
         </div>
         <div class="right">
-            <input type="text" class="text" name="episodes" title="Episodes filter. Please read the instructions on how to use this feature." value="<?php echo $item['Episodes'] ?>">
+            <input type="text" class="text" name="episodes" id="favorite_<?php echo $key; ?>_episodes" title="Episodes filter. Please read the instructions on how to use this feature." value="<?php echo $item['Episodes'] ?>">
         </div>
     </div>
     <div class="favorite_feed">
@@ -60,7 +60,7 @@ if (isset($style)) {
             <label class="item" title="Feed to match against">Feed:</label>
         </div>
         <div class="right">
-            <select name="feed">
+            <select name="feed" id="favorite_<?php echo $key; ?>_feed">
                 <?php echo $feed_options; ?>
             </select>
         </div>
@@ -70,7 +70,7 @@ if (isset($style)) {
             <label class="item" title="Search for this quality in the full title">Quality:</label>
         </div>
         <div class="right">
-            <input type="text" class="text" name="quality" value="<?php echo $item['Quality']; ?>">
+            <input type="text" class="text" name="quality" id="favorite_<?php echo $key ?>_quality" value="<?php echo $item['Quality']; ?>">
         </div>
     </div>
     <div class="favorite_seed_and_episode">
@@ -86,18 +86,18 @@ if (isset($style)) {
             if ($config_values['Settings']['Client'] == "folder") {
                 echo 'style="display: none"';
             }
-            ?> name="seedratio" value="<?php echo getArrayValueByKey($item, 'seedRatio'); ?>">
+            ?> name="seedratio" id="favorite_<?php echo $key; ?>_seedratio" value="<?php echo getArrayValueByKey($item, 'seedRatio'); ?>">
             <label class="lastSeason item" title="SSxEE or YYYYMMDD notation only" <?php
             if ($config_values['Settings']['Client'] == "folder") {
                 echo 'style="padding-left: 137px"';
             }
             ?>>Last Download:</label>
                    <?php if (isset($item['Episode']) && $item['Season'] !== 0 && !preg_match('/^(\d{8})$/', $item['Episode'])) { ?>
-                <input class='lastSeason text' type="text" name="season" value="<?php echo $item['Season']; ?>">
+                <input class='lastSeason text' type="text" name="season" id="favorite_<?php echo $key; ?>_season" value="<?php echo $item['Season']; ?>">
                 <label class="lastEpisode item">x</label>
-                <input class='lastEpisode text' type="text" name="episode" value="<?php echo $item['Episode']; ?>">
+                <input class='lastEpisode text' type="text" name="episode" id="favorite_<?php echo $key; ?>_episode" value="<?php echo $item['Episode']; ?>">
             <?php } else { ?>
-                <input class='lastEpisode text' type="text" name="episode" value="<?php
+                <input class='lastEpisode text' type="text" name="episode" id="favorite_<?php echo $key; ?>_episode" value="<?php
                 if (isset($item['Episode'])) {
                     echo $item['Episode'];
                 }
@@ -106,7 +106,9 @@ if (isset($style)) {
         </div>
     </div>
     <div class="buttonContainer">
-        <a class="submitForm button" id="Update" href="#">Update</a>
-        <a class="submitForm button" id="Delete" href="#favorite_<?php echo $key ?>">Delete</a>
+        <!--<a class="submitForm button" id="favorite_<?php //echo $key ?>_Update" href="#">Update</a>
+        <a class="submitForm button" id="favorite_<?php //echo $key ?>_Delete" href="#favorite_<?php //echo $key  ?>">Delete</a>-->
+        <a class="submitForm button" id="Update" href="#favorite_<?php echo $key  ?>">Update</a>
+        <a class="submitForm button" id="Delete" href="#favorite_<?php echo $key  ?>">Delete</a>
     </div>
 </form>